+++ internal/pkg/server/prepareRename.go
@@ -1,0 +1,21 @@
+package server
+
+import (
+	"github.com/docker/docker-language-server/internal/compose"
+	"github.com/docker/docker-language-server/internal/pkg/document"
+	"github.com/docker/docker-language-server/internal/tliron/glsp"
+	"github.com/docker/docker-language-server/internal/tliron/glsp/protocol"
+	"go.lsp.dev/uri"
+)
+
+func (s *Server) TextDocumentPrepareRename(ctx *glsp.Context, params *protocol.PrepareRenameParams) (any, error) {
+	doc, err := s.docs.Read(ctx.Context, uri.URI(params.TextDocument.URI))
+	if err != nil {
+		return nil, err
+	}
+	defer doc.Close()
+	if doc.LanguageIdentifier() == protocol.DockerComposeLanguage {
+		return compose.PrepareRename(doc.(document.ComposeDocument), params)
+	}
+	return nil, nil
+}